move AU_solar into subdirectory - restores commandline UI
Also, don't try to be so clever sorting datetimes.
